server/src/requirements.txt
pyramid==1.10.5
mysql-connector-python
python-dotenv